streamlit/requirements.txt
streamlit
pandas
numpy
scikit-learn==1.4.2
joblibpip uninstall imbalanced-learn -y